+++ src/main/java/com/bookspot/batch/step/processor/csv/stock/repository/BookRepository.java
@@ -1,4 +1,4 @@
-package com.bookspot.batch.library;
+package com.bookspot.batch.step.processor.csv.stock.repository;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -9,16 +9,16 @@ import java.util.Optional;
 
 @Repository
 @RequiredArgsConstructor
-public class LibraryRepository {
+public class BookRepository {
     private final JdbcTemplate jdbcTemplate;
-    private static final String LIBRARY_CODE_QUERY = "SELECT id FROM Library WHERE library_code = ?";
+    private static final String ISBN_QUERY = "SELECT id FROM Book WHERE isbn13 = ?";
 
-    public Optional<Long> findId(String libraryCode) {
+    public Optional<Long> findIdByIsbn13(String isbn13) {
         try {
             return Optional.of(jdbcTemplate
                     .queryForObject(
-                            LIBRARY_CODE_QUERY,
-                            new Object[]{libraryCode},
+                            ISBN_QUERY,
+                            new Object[]{isbn13},
                             Long.class));
         } catch (EmptyResultDataAccessException e) {
             return Optional.empty();
